--- IOSassignment-main-2/README.md
@@ -1,1 +1,0 @@
-Github Link: https://github.com/tomikode/IOSassignment
