--- IOSassignment-main-2/Restaurant Booking app.xcodeproj/xcuserdata/giaphathuynh.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "BB50B72D-882A-452F-B3DC-B26C29B7AB1C"
-   type = "1"
-   version = "2.0">
-</Bucket>
