--- IOSassignment-main-2/Restaurant Booking app/Assets.xcassets/Risotto.imageset/Contents.json
@@ -1,21 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "Risotto.jpeg",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
