--- IOSassignment-main-2/Restaurant Booking app/MenuTableViewCell.swift
@@ -1,28 +1,0 @@
-//
-//  MenuTableViewCell.swift
-//  Restaurant Booking app
-//
-//  Created by Gia Phat Huynh on 18/5/22.
-//
-
-import UIKit
-
-class MenuTableViewCell: UITableViewCell {
-
-    @IBOutlet var myLabel: UILabel!
-    @IBOutlet var myImageView: UIImageView!
-    @IBOutlet var descripLabel: UILabel!
-    @IBOutlet var priceLabel: UILabel!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-    
-}
