--- IOSassignment-main-2/Restaurant Booking app/MenuTableViewCell.xib
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="19529" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19519"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" reuseIdentifier="MenuTableViewCell" rowHeight="310" id="KGk-i7-Jjw" customClass="MenuTableViewCell" customModule="Restaurant_Booking_app" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="251" height="310"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="251" height="310"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="Pizza" translatesAutoresizingMaskIntoConstraints="NO" id="MXC-LK-1aT">
-                        <rect key="frame" x="10" y="10" width="231" height="183"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="183" id="P53-dW-ojg"/>
-                        </constraints>
-                    </imageView>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="252" text="Description" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="GbE-Yy-UNF">
-                        <rect key="frame" x="26" y="253" width="205" height="47"/>
-                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Pizza" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="u3a-uL-G5t">
-                        <rect key="frame" x="26" y="208" width="199" height="30"/>
-                        <fontDescription key="fontDescription" style="UICTFontTextStyleTitle1"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="$24" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="q4W-yb-TGG">
-                        <rect key="frame" x="195" y="218" width="30" height="20.5"/>
-                        <fontDescription key="fontDescription" type="system" weight="light" pointSize="17"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                </subviews>
-                <constraints>
-                    <constraint firstItem="MXC-LK-1aT" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="10" id="0BP-kO-W0P"/>
-                    <constraint firstItem="u3a-uL-G5t" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leadingMargin" constant="10" id="1d5-VS-UK9"/>
-                    <constraint firstAttribute="trailingMargin" secondItem="q4W-yb-TGG" secondAttribute="trailing" constant="10" id="34C-sr-Ljb"/>
-                    <constraint firstAttribute="trailing" secondItem="MXC-LK-1aT" secondAttribute="trailing" constant="10" id="5F3-Ml-o2F"/>
-                    <constraint firstAttribute="bottom" secondItem="GbE-Yy-UNF" secondAttribute="bottom" constant="10" id="C9j-yj-JMD"/>
-                    <constraint firstAttribute="trailing" secondItem="GbE-Yy-UNF" secondAttribute="trailing" constant="20" id="EPF-AQ-lml"/>
-                    <constraint firstItem="MXC-LK-1aT" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" constant="10" id="Kba-uP-gvw"/>
-                    <constraint firstItem="q4W-yb-TGG" firstAttribute="top" secondItem="MXC-LK-1aT" secondAttribute="bottom" constant="25" id="LQq-EC-7R6"/>
-                    <constraint firstAttribute="trailingMargin" secondItem="u3a-uL-G5t" secondAttribute="trailing" constant="10" id="Nev-SP-Yi2"/>
-                    <constraint firstItem="u3a-uL-G5t" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leadingMargin" constant="10" id="ReN-jm-rlR"/>
-                    <constraint firstItem="u3a-uL-G5t" firstAttribute="top" secondItem="MXC-LK-1aT" secondAttribute="bottom" constant="15" id="XxB-z6-bBG"/>
-                    <constraint firstItem="GbE-Yy-UNF" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="26" id="pgF-b3-fbQ"/>
-                    <constraint firstItem="GbE-Yy-UNF" firstAttribute="top" secondItem="u3a-uL-G5t" secondAttribute="bottom" constant="15" id="sM9-gV-f3N"/>
-                    <constraint firstItem="GbE-Yy-UNF" firstAttribute="top" secondItem="MXC-LK-1aT" secondAttribute="bottom" constant="60" id="xcK-Ib-dCM"/>
-                    <constraint firstItem="GbE-Yy-UNF" firstAttribute="leading" secondItem="u3a-uL-G5t" secondAttribute="leading" id="yG6-uX-U2M"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <connections>
-                <outlet property="descripLabel" destination="GbE-Yy-UNF" id="oLj-7y-NNJ"/>
-                <outlet property="myImageView" destination="MXC-LK-1aT" id="j67-cE-mkY"/>
-                <outlet property="myLabel" destination="u3a-uL-G5t" id="No4-Xu-fYw"/>
-                <outlet property="priceLabel" destination="q4W-yb-TGG" id="dEs-By-Ygv"/>
-            </connections>
-            <point key="canvasLocation" x="81.884057971014499" y="64.955357142857139"/>
-        </tableViewCell>
-    </objects>
-    <resources>
-        <image name="Pizza" width="275" height="183"/>
-    </resources>
-</document>
