--- IOSassignment-main-2/Restaurant Booking app/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Restaurant Booking app
-//
-//  Created by Tomi Piisila on 1/5/2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
